Use static_cast to silence conversion warning in ListBoxBase
We should probably consider using `int` for those values.

+++ OPHD/UI/Core/ListBoxBase.h
@@ -80,8 +80,8 @@ protected:
 
 	void _update_item_display();
 
-	unsigned int item_width() const { return mItemWidth; }
-	unsigned int item_height() const { return mItemHeight; }
+	unsigned int item_width() const { return static_cast<unsigned int>(mItemWidth); }
+	unsigned int item_height() const { return static_cast<unsigned int>(mItemHeight); }
 	void item_height(int);
 
 	unsigned int draw_offset() const { return mCurrentOffset; }
